refactor: move any worker related stuff out of service module
