+++ src/starlite_saqlalchemy/service/__init__.py
@@ -1,8 +1,5 @@
 """Implementations for service object."""
 
-from .generic import Service, make_service_callback
+from .generic import Service
 
-__all__ = [
-    "Service",
-    "make_service_callback",
-]
+__all__ = ["Service"]
