+++ tests/unit/test_service.py
@@ -7,18 +7,14 @@ import pytest
 pytest.importorskip("sqlalchemy")
 
 from datetime import date
-from typing import TYPE_CHECKING, Any
-from unittest.mock import AsyncMock
+from typing import TYPE_CHECKING
 from uuid import uuid4
 
-from saq import Job
-
-from starlite_saqlalchemy import db, service, worker
+from starlite_saqlalchemy import service
 from starlite_saqlalchemy.exceptions import NotFoundError
 from tests.utils import domain
 
 if TYPE_CHECKING:
-    from pytest import MonkeyPatch
 
     from starlite_saqlalchemy.testing.generic_mock_repository import (
         GenericMockRepository,
@@ -90,71 +86,6 @@ async def test_service_delete() -> None:
     author, _ = await service_obj.list()
     deleted = await service_obj.delete(author.id)
     assert author is deleted
-
-
-async def test_make_service_callback(
-    raw_authors: list[dict[str, Any]], monkeypatch: "MonkeyPatch"
-) -> None:
-    """Tests loading and retrieval of service object types."""
-    recv_cb_mock = AsyncMock()
-    monkeypatch.setattr(service.Service, "receive_callback", recv_cb_mock, raising=False)
-    await service.make_service_callback(
-        {},
-        service_type_id="tests.utils.domain.authors.Service",
-        service_method_name="receive_callback",
-        raw_obj=raw_authors[0],
-    )
-    recv_cb_mock.assert_called_once_with(raw_obj=raw_authors[0])
-
-
-async def test_make_service_callback_raises_runtime_error(
-    raw_authors: list[dict[str, Any]]
-) -> None:
-    """Tests loading and retrieval of service object types."""
-    with pytest.raises(KeyError):
-        await service.make_service_callback(
-            {},
-            service_type_id="tests.utils.domain.LSKDFJ",
-            service_method_name="receive_callback",
-            raw_obj=raw_authors[0],
-        )
-
-
-async def test_enqueue_service_callback(monkeypatch: "MonkeyPatch") -> None:
-    """Tests that job enqueued with desired arguments."""
-    enqueue_mock = AsyncMock()
-    monkeypatch.setattr(worker.queue, "enqueue", enqueue_mock)
-    service_instance = domain.authors.Service(session=db.async_session_factory())
-    await service_instance.enqueue_background_task("receive_callback", raw_obj={"a": "b"})
-    enqueue_mock.assert_called_once()
-    assert isinstance(enqueue_mock.mock_calls[0].args[0], Job)
-    job = enqueue_mock.mock_calls[0].args[0]
-    assert job.function == service.make_service_callback.__qualname__
-    assert job.kwargs == {
-        "service_type_id": "tests.utils.domain.authors.Service",
-        "service_method_name": "receive_callback",
-        "raw_obj": {"a": "b"},
-    }
-
-
-async def test_enqueue_service_callback_with_custom_job_config(monkeypatch: "MonkeyPatch") -> None:
-    """Tests that job enqueued with desired arguments."""
-    enqueue_mock = AsyncMock()
-    monkeypatch.setattr(worker.queue, "enqueue", enqueue_mock)
-    service_instance = domain.authors.Service(session=db.async_session_factory())
-    await service_instance.enqueue_background_task(
-        "receive_callback", job_config=worker.JobConfig(timeout=999), raw_obj={"a": "b"}
-    )
-    enqueue_mock.assert_called_once()
-    assert isinstance(enqueue_mock.mock_calls[0].args[0], Job)
-    job = enqueue_mock.mock_calls[0].args[0]
-    assert job.function == service.make_service_callback.__qualname__
-    assert job.timeout == 999
-    assert job.kwargs == {
-        "service_type_id": "tests.utils.domain.authors.Service",
-        "service_method_name": "receive_callback",
-        "raw_obj": {"a": "b"},
-    }
 
 
 async def test_service_new_context_manager() -> None:
